requirements.txt
streamlit
PyPDF2
sentence-transformers
openai
torch
python-dotenv